+++ config.json
@@ -1,40 +1,58 @@
 {
-  "model" : {
-    "min_input_size":       288,
-    "max_input_size":       448,
-    "anchors":              [33,34, 52,218, 55,67, 92,306, 96,88, 118,158, 153,347, 209,182, 266,359],
-    "labels":               ["helmet","person with helmet","person without helmet"]
+  "model": {
+    "min_input_size": 288,
+    "max_input_size": 448,
+    "anchors": [
+      33,
+      34,
+      52,
+      218,
+      55,
+      67,
+      92,
+      306,
+      96,
+      88,
+      118,
+      158,
+      153,
+      347,
+      209,
+      182,
+      266,
+      359
+    ],
+    "labels": ["capacete", "pessoa com capacete", "pessoa sem capacete"]
   },
 
   "train": {
-    "train_image_folder":   "train_image_folder/",
-    "train_annot_folder":   "train_annot_folder/",
-    "cache_name":           "helmet_train.pkl",
+    "train_image_folder": "train_image_folder/",
+    "train_annot_folder": "train_annot_folder/",
+    "cache_name": "helmet_train.pkl",
 
-    "train_times":          8,
-    "batch_size":           8,
-    "learning_rate":        1e-4,
-    "nb_epochs":            100,
-    "warmup_epochs":        3,
-    "ignore_thresh":        0.5,
-    "gpus":                 "0,1",
+    "train_times": 8,
+    "batch_size": 8,
+    "learning_rate": 1e-4,
+    "nb_epochs": 100,
+    "warmup_epochs": 3,
+    "ignore_thresh": 0.5,
+    "gpus": "0,1",
 
-    "grid_scales":          [1,1,1],
-    "obj_scale":            5,
-    "noobj_scale":          1,
-    "xywh_scale":           1,
+    "grid_scales": [1, 1, 1],
+    "obj_scale": 5,
+    "noobj_scale": 1,
+    "xywh_scale": 1,
 
-
-    "tensorboard_dir":      "logs",
-    "saved_weights_name":   "full_yolo3_helmet_and_person.h5",
-    "debug":                true
+    "tensorboard_dir": "logs",
+    "saved_weights_name": "full_yolo3_helmet_and_person.h5",
+    "debug": true
   },
 
   "valid": {
-    "valid_image_folder":   "",
-    "valid_annot_folder":   "",
-    "cache_name":           "",
+    "valid_image_folder": "",
+    "valid_annot_folder": "",
+    "cache_name": "",
 
-    "valid_times":          1
+    "valid_times": 1
   }
 }
