--- links.txt
@@ -1,3 +1,0 @@
-mars-small128.pb : https://1drv.ms/u/s!ArJHK_Eldk0Cg3jyt-NR3xPErr_5?e=88vcgg
-full_yolo3_helmet_and_person.h5 : https://1drv.ms/u/s!ArJHK_Eldk0Cg3cTEpkVoZyyxQzl?e=10MXuV
-backend.h5 : https://1drv.ms/u/s!ArJHK_Eldk0Cg3nUkkHZcS7btEGb?e=BlFGvM
