+++ voc.py
